up for minecraft 1.6.2

+++ entity/EntityDoubleLavaBoat.java
@@ -44,7 +44,7 @@ public class EntityDoubleLavaBoat extends EntityDoubleBoat {
      * Called when the entity is attacked.
      */
     @Override
-    public boolean attackEntityFrom(DamageSource damageSource, int par2) {
+    public boolean attackEntityFrom(DamageSource damageSource, float par2) {
         return attackEntityFrom(damageSource, 4, par2);
     }
