+++ entity/EntityDoubleReinforcedBoat.java
@@ -34,7 +34,7 @@ public class EntityDoubleReinforcedBoat extends EntityDoubleBoat {
      * Called when the entity is attacked.
      */
     @Override
-    public boolean attackEntityFrom(DamageSource damageSource, int par2) {
+    public boolean attackEntityFrom(DamageSource damageSource, float par2) {
         return attackEntityFrom(damageSource, 1, par2);
     }
