+++ entity/EntityLavaBoat.java
@@ -44,7 +44,7 @@ public class EntityLavaBoat extends EntityNKBoat {
      * Called when the entity is attacked.
      */
     @Override
-    public boolean attackEntityFrom(DamageSource damageSource, int par2) {
+    public boolean attackEntityFrom(DamageSource damageSource, float par2) {
         return attackEntityFrom(damageSource, 3, par2);
     }
